Separate classes into files
